+++ src/otp.cpp
@@ -2,6 +2,7 @@
 
 
 #include "internal_error.h"
+#include "action_error.h"
 #include "global_data.h"
 #include "logger.h"
 
@@ -75,29 +76,6 @@ otp [-D | -F | -DF ] [action_name] [action params]
 
 
 */
-
-
-
-class action_error : public std::exception {
-	std::string message;
-public:
-	static void assert_true(bool condition, const action_error& exception) {
-		if (!condition) throw exception;
-	}
-
-	action_error(const std::string& message) : message(message) {}
-	action_error(const char* message) : action_error(std::string(message)) {}
-	action_error(const action_error&) = default;
-
-
-
-	const char* what() const noexcept override {
-		return message.c_str();
-	}
-};
-
-
-
 
 
 class action {
